Close Connection Handshake and Backpressure Adjustment
It is important that the socket is not closed directly after sending all tuples, as it is possible that some of them will get dropped, since the connection was close. The Close Connection handshake enforces that the connection stays open until the Unikernel acknowledged that DONE packt.

Backpressure Work similar, if the testbench encounters too much blocking we assume that the unikernel might be under to heavy load, in this case we drop back to a blocking socket and wait until the BACK package gets acknowledged.